+++ src/features/notifications/components/NotificationItem/NotificationItem.tsx
@@ -3,7 +3,7 @@
 import Link from 'next/link';
 import Image from 'next/image';
 import { NotificationWithRelations } from '@/features/notifications/types';
-import { timeAgo } from '@/utils';
+import { timeAgo } from '@/shared/utils/date';
 import { AiFillHeart } from 'react-icons/ai';
 import { FaComment } from 'react-icons/fa';
