+++ src/shared/utils/string.ts
