--- src/utils/index.ts
@@ -1,2 +1,0 @@
-export * from './date';
-export * from './string';
